requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
fastapi==0.110.0
h11==0.14.0
idna==3.6
imageio==2.34.2
imageio-ffmpeg==0.5.1
kombu==5.3.7
moviepy==1.0.3
numpy==2.0.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-multipart==0.0.9
redis==5.0.7
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
loguru==0.7.2
